migrate paths
