+++ app/database_utils.py
@@ -1,11 +1,7 @@
 from sqlalchemy import create_engine
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import sessionmaker
-
-import psycopg2
-from psycopg2.extras import DictCursor, RealDictCursor
 import os
-import time
 from dotenv import load_dotenv
 
 load_dotenv()
@@ -31,24 +27,3 @@ def get_db():
         yield db
     finally:
         db.close()
-
-
-def db_connection():
-    while True:
-        try:
-            connection = psycopg2.connect(
-                host=os.getenv("HOST"),
-                database=os.getenv("DATABASE"),
-                user=os.getenv("USER"),
-                password=os.getenv("PASSWORD"),
-                port=os.getenv("PORT"),
-            )
-            cursor_dict = connection.cursor(cursor_factory=DictCursor)
-            cursor_realdict = connection.cursor(cursor_factory=RealDictCursor)
-            # cursor = connection.cursor()
-            print("Database connection was succeful \U0001F44C !")
-            return connection, cursor_dict, cursor_realdict
-        except (Exception, psycopg2.Error) as error:
-            print("Connecting to Database failed...")
-            print("Error : ", error)
-            time.sleep(2)
